refactor: add utils file
