+++ crates/runnair/src/main.rs
@@ -1,6 +1,7 @@
 use vm::run_fibonacci;
 
 pub mod memory;
+pub mod utils;
 pub mod vm;
 
 fn main() {
